:recycle: Rename registerV2 to register
