+++ src/components/register/accountCreated.css
