+++ src/components/register/backupPassphrase.js
@@ -3,7 +3,7 @@ import React from 'react';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import { PrimaryButtonV2, TertiaryButtonV2 } from '../toolbox/buttons/button';
 import PassphraseBackup from '../passphraseBackup';
-import registerStyles from './registerV2.css';
+import registerStyles from './register.css';
 
 const BackupPassphrase = ({
   t, account, prevStep, nextStep,
