+++ src/components/register/backupPassphrase.test.js
