+++ src/components/register/chooseAvatar.css
