+++ src/components/register/chooseAvatar.test.js
