+++ src/components/register/confirmPassphrase.css
