+++ src/components/register/confirmPassphrase.test.js
