+++ src/components/register/confirmPassphraseOptions.js
