+++ src/components/register/confirmPassphraseOptions.test.js
