+++ src/components/splashscreen/splashscreen.js
@@ -89,7 +89,7 @@ class Splashscreen extends React.Component {
             <Link className={`${styles.button} login-button`} to={routes.loginV2.path}>
               <SecondaryButtonV2 className={'light'}>{t('Sign In')}</SecondaryButtonV2>
             </Link>
-            <Link className={`${styles.button} new-account-button`} to={routes.registerV2.path}>
+            <Link className={`${styles.button} new-account-button`} to={routes.register.path}>
               <PrimaryButtonV2>{t('Create an Account')}</PrimaryButtonV2>
             </Link>
             <span className={styles.separator}>
